LayoutTests/imported/w3c/web-platform-tests/css/css-text-decor/text-decoration-skip-ink-expected.txt


FAIL Body must have text-decoration-skip-ink auto by default. assert_equals: Must be set to value auto as initial value. expected (string) "auto" but got (undefined) undefined
FAIL Property text-decoration-skip-ink must support values auto, none and all. assert_true: Text-decoration-skip-ink must allow setting the value auto. expected true got false
